backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy>=2.0.33
psycopg[binary]==3.2.12
python-dotenv==1.0.0
requests==2.31.0
pydantic<2
python-multipart==0.0.6